+++ frontend/src/ui/login/SignIn.module.css
@@ -1,0 +1,35 @@
+
+.icon {
+    width: 30px; /* Adjust icon size as needed */
+    margin-right: 10px; /* Space between icon and button */
+    margin-top: 5px;
+    height: 30px; /* Ensures the aspect ratio is maintained */
+    float: left;
+
+}
+  
+.button {
+    padding: 10px 25px; /* Adjusted padding for better spacing */
+    color: white; /* Text color */
+    background-color: #B28A68; /* Background color */
+    border: none; /* Removed border */
+    border-radius: 5px; /* Rounded corners */
+    font-size: 25px; /* Font size */
+    font-weight: bolder; /* Thicker font */
+    font-style: normal;
+    cursor: pointer; /* Pointer cursor on hover */
+    transition: background-color 0.3s ease; /* Smooth transition for background color change */
+    margin-top: 200px;
+    margin-bottom: 20px;
+    margin-left: 600px;
+    display: block; /* Display as block to center horizontally */
+}
+
+.container {
+    color:white;
+    width: 300px; /* Adjust as needed */
+    padding: 20px; /* Adjust as needed */
+    margin: 0 auto; /* Center the container */
+    border: 1px solid #ccc; /* Add a border */
+    border-radius: 5px; /* Optional: add rounded corners */
+}
